refactor: Add tenant_id variable in session functions
