+++ supertokens_python/recipe/multitenancy/recipe.py
@@ -259,41 +259,43 @@ class APIImplementation(APIInterface):
 
 class AllowedDomainsClaimClass(PrimitiveArrayClaim[List[str]]):
     def __init__(self):
-        async def fetch_value(_user_id: str, user_context: Dict[str, Any]) -> List[str]:
+        default_max_age_in_sec = 60 * 60 * 24 * 7
+
+        async def fetch_value(
+            _: str, tenant_id: str, user_context: Dict[str, Any]
+        ) -> Optional[List[str]]:
             recipe = MultitenancyRecipe.get_instance()
-            tenant_id = (
-                None  # TODO fetch value will be passed with tenant_id as well later
-            )
 
-            if recipe.config.get_allowed_domains_for_tenant_id is None:
-                return (
-                    []
-                )  # User did not provide a function to get allowed domains, but is using a validator. So we don't allow any domains by default
+            if recipe.get_allowed_domains_for_tenant_id is None:
+                # User did not provide a function to get allowed domains, but is using a validator. So we don't allow any domains by default
+                return None
 
-            domains_res = await recipe.config.get_allowed_domains_for_tenant_id(
+            return await recipe.get_allowed_domains_for_tenant_id(
                 tenant_id, user_context
             )
-            return domains_res
 
-        super().__init__(
-            key="st-tenant-domains",
-            fetch_value=fetch_value,
-            default_max_age_in_sec=3600,
-        )
+        super().__init__("st-t-dmns", fetch_value, default_max_age_in_sec)
 
     def get_value_from_payload(
-        self, payload: JSONObject, user_context: Union[Dict[str, Any], None] = None
-    ) -> Union[List[str], None]:
-        if self.key not in payload:
+        self, payload: JSONObject, user_context: Optional[Dict[str, Any]] = None
+    ) -> Optional[List[str]]:
+        _ = user_context
+
+        res = payload.get(self.key, {}).get("v")
+        if res is None:
             return []
-        return super().get_value_from_payload(payload, user_context)
+        return res
 
     def get_last_refetch_time(
-        self, payload: JSONObject, user_context: Union[Dict[str, Any], None] = None
-    ) -> Union[int, None]:
-        if self.key not in payload:
+        self, payload: JSONObject, user_context: Optional[Dict[str, Any]] = None
+    ) -> Optional[int]:
+        _ = user_context
+
+        res = payload.get(self.key, {}).get("t")
+        if res is None:
             return get_timestamp_ms()
-        return super().get_last_refetch_time(payload, user_context)
+
+        return res
 
 
 AllowedDomainsClaim = AllowedDomainsClaimClass()
