+++ supertokens_python/recipe/session/claim_base_classes/boolean_claim.py
@@ -31,7 +31,7 @@ class BooleanClaim(PrimitiveClaim[bool]):
         self,
         key: str,
         fetch_value: Callable[
-            [str, Dict[str, Any]],
+            [str, str, Dict[str, Any]],
             MaybeAwaitable[Optional[bool]],
         ],
         default_max_age_in_sec: Optional[int] = None,
