+++ supertokens_python/recipe/session/session_request_functions.py
@@ -238,7 +238,7 @@ async def create_new_session_in_request(
     final_access_token_payload = {**access_token_payload, "iss": issuer}
 
     for claim in claims_added_by_other_recipes:
-        update = await claim.build(user_id, user_context)
+        update = await claim.build(user_id, tenant_id, user_context)
         final_access_token_payload = {**final_access_token_payload, **update}
 
     log_debug_message("createNewSession: Access token payload built")
